Do (flowcell) retention in PCR step
Previously was done by the controller after PCR
However PCR generates exponential numbers of copies
most of which are discarded immediately. Instead, we generate
only those copies that are retained, while still preserving the full
'binary tree' behavior of PCR copying which will be important for
PCR errors and dupes.