Re-arrange config to reduce Snakemake rerunning stages
Previously all config options were passed to all parts of BEERS,
so a change in any config option required a complete re-run of BEERS.
Now, everything gets a 'global' section, so changes to that will force
a rerun from the top. However, changes to library prep or sequence
pipeline specific config will not have to rerun all of the rest.